bugfix: icon src in feeds
Icons weren't showing up in feeds because I forgot to change the src of those icons. To help prevent this in the future, I used our template for outputting the icon source (and made the icon source point to the absolute location of the file instead of relative, i.e. `http://.../img` instead of `/img`)