refactor: display of 256, 512, and 1024 icons
The watchOS icons gallery used to not have 512 source images on some post, which resulted in some tricky logic around what and how to display. I went in and upsized those old images from about ~300px to 512px so we can guarantee all posts have at least a 512px image (this seems fine, as we were already letting the browser upsize a 300px to 512px at display time anyway).

This simplifies a lot of the display logic around icons. It also makes the build process just slightly simpler. It's nice to know we can guarantee a 512px image for all icons.